requirements.txt
annotated-types==0.7.0
av==12.3.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cmake==3.31.0.1
colorama==0.4.6
coloredlogs==15.0.1
ctranslate2==4.5.0
dlib==19.24.6
faster-whisper==1.0.3
filelock==3.16.1
Flask==3.0.3
flatbuffers==24.3.25
fpdf==1.7.2
fsspec==2024.10.0
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.151.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
grpcio==1.67.1
grpcio-status==1.67.1
httplib2==0.22.0
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mpmath==1.3.0
numpy==2.1.3
onnxruntime==1.20.0
opencv-python==4.10.0.84
packaging==24.2
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
pyreadline3==3.5.4
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
sympy==1.13.3
tokenizers==0.20.3
tqdm==4.67.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
Werkzeug==3.1.3
